collector/requirements.txt
google-api-python-client>=2.100.0
httpx>=0.27.0
python-dotenv>=1.0.0
thefuzz[speedup]>=0.22.0
pyyaml>=6.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
